Move restoreDataFromTimestamp into data.go
Grouping the data related functions in the same file will make it easier
to work on them. This is now more similar to how backup is organized.
